feat: inline-editable assets in Financial data (#45)
Make Investment/Tangible asset/Liability/Income/Expense items in the
Financial data section editable in place instead of static read-only
cards. Extract the per-category editing logic that the onboarding
finances/edit/* pages duplicated into five reusable *-editor components,
now shared by both onboarding and financial-data.

- Cards are collapsed by default and expand (ChevronsUpDown) into the
  same editable form as onboarding.
- Edits (and add/duplicate/delete) auto-save to the profile immediately;
  the donut/total update live.
- The top-right Update button resumes the onboarding flow at the current
  category; the back arrow always returns to Home in one press.
- Remove the now-unused read-only card components and orphaned locale keys.

Closes #45
