+++ src/routes/(app)/financial-data/investments/+page.svelte
@@ -1,16 +1,10 @@
 <script lang="ts">
   import { _, locale } from 'svelte-i18n'
 
-  import { Plus } from '@lucide/svelte'
-
-  import { resolve } from '$app/paths'
-
   import { CATEGORY_COLORS } from '$lib/chart-colors'
   import DonutChart from '$lib/components/donut-chart.svelte'
-  import ReadOnlyItemCard from '$lib/components/read-only-item-card.svelte'
-  import { Button } from '$lib/components/ui/button'
+  import InvestmentsEditor from '$lib/components/investments-editor.svelte'
   import { getInvestmentsTotal } from '$lib/financial-totals'
-  import routes from '$lib/routes'
   import { appStore } from '$lib/stores/app.svelte'
   import { formatLastUpdated } from '$lib/utils'
 
@@ -45,23 +39,4 @@
   {$_('page.financialData.investments.description')}
 </p>
 
-{#if investments.length === 0}
-  <p class="w-full py-8 text-center text-sm text-muted-foreground">
-    {$_('page.financialData.investments.empty')}
-  </p>
-{:else}
-  <div class="flex w-full flex-col gap-2">
-    {#each investments as investment, idx (investment.id)}
-      <ReadOnlyItemCard
-        name={investment.name}
-        value={appStore.formatCurrencyCode(investment.balance)}
-        dotColor={CATEGORY_COLORS.investments[idx % CATEGORY_COLORS.investments.length]}
-      />
-    {/each}
-  </div>
-{/if}
-
-<Button variant="secondary" size="sm" href={resolve(routes.FINANCES_EDIT_INVESTMENTS)}>
-  <Plus class="size-4" />
-  {$_('page.setup.investments.addInvestment')}
-</Button>
+<InvestmentsEditor />
