+++ src/routes/(app)/financial-data/liabilities/+page.svelte
@@ -1,16 +1,10 @@
 <script lang="ts">
   import { _, locale } from 'svelte-i18n'
 
-  import { Plus } from '@lucide/svelte'
-
-  import { resolve } from '$app/paths'
-
   import { CATEGORY_COLORS } from '$lib/chart-colors'
   import DonutChart from '$lib/components/donut-chart.svelte'
-  import ReadOnlyItemCard from '$lib/components/read-only-item-card.svelte'
-  import { Button } from '$lib/components/ui/button'
+  import LiabilitiesEditor from '$lib/components/liabilities-editor.svelte'
   import { getLiabilitiesTotal } from '$lib/financial-totals'
-  import routes from '$lib/routes'
   import { appStore } from '$lib/stores/app.svelte'
   import { formatLastUpdated } from '$lib/utils'
 
@@ -47,24 +41,4 @@
   {$_('page.financialData.liabilities.description')}
 </p>
 
-{#if liabilities.length === 0}
-  <p class="w-full py-8 text-center text-sm text-muted-foreground">
-    {$_('page.financialData.liabilities.empty')}
-  </p>
-{:else}
-  <div class="flex w-full flex-col gap-2">
-    {#each liabilities as liability, idx (liability.id)}
-      <ReadOnlyItemCard
-        name={liability.name}
-        value={appStore.formatCurrencyCode(-liability.outstanding_balance)}
-        valueClass="text-destructive"
-        dotColor={CATEGORY_COLORS.liabilities[idx % CATEGORY_COLORS.liabilities.length]}
-      />
-    {/each}
-  </div>
-{/if}
-
-<Button variant="secondary" size="sm" href={resolve(routes.FINANCES_EDIT_LIABILITIES)}>
-  <Plus class="size-4" />
-  {$_('page.setup.liabilities.addLiability')}
-</Button>
+<LiabilitiesEditor />
