+++ src/routes/(app)/financial-data/tangible-assets/+page.svelte
@@ -1,16 +1,10 @@
 <script lang="ts">
   import { _, locale } from 'svelte-i18n'
 
-  import { Plus } from '@lucide/svelte'
-
-  import { resolve } from '$app/paths'
-
   import { CATEGORY_COLORS } from '$lib/chart-colors'
   import DonutChart from '$lib/components/donut-chart.svelte'
-  import ReadOnlyItemCard from '$lib/components/read-only-item-card.svelte'
-  import { Button } from '$lib/components/ui/button'
+  import TangibleAssetsEditor from '$lib/components/tangible-assets-editor.svelte'
   import { getTangibleAssetsTotal } from '$lib/financial-totals'
-  import routes from '$lib/routes'
   import { appStore } from '$lib/stores/app.svelte'
   import { formatLastUpdated } from '$lib/utils'
 
@@ -45,23 +39,4 @@
   {$_('page.financialData.tangibleAssets.description')}
 </p>
 
-{#if tangibleAssets.length === 0}
-  <p class="w-full py-8 text-center text-sm text-muted-foreground">
-    {$_('page.financialData.tangibleAssets.empty')}
-  </p>
-{:else}
-  <div class="flex w-full flex-col gap-2">
-    {#each tangibleAssets as asset, idx (asset.id)}
-      <ReadOnlyItemCard
-        name={asset.name}
-        value={appStore.formatCurrencyCode(asset.value)}
-        dotColor={CATEGORY_COLORS.tangibleAssets[idx % CATEGORY_COLORS.tangibleAssets.length]}
-      />
-    {/each}
-  </div>
-{/if}
-
-<Button variant="secondary" size="sm" href={resolve(routes.FINANCES_EDIT_TANGIBLE_ASSETS)}>
-  <Plus class="size-4" />
-  {$_('page.setup.tangibleAssets.addAsset')}
-</Button>
+<TangibleAssetsEditor />
